Refactor user settings management; rename save_user_settings to apply_user_settings and update related references; add new GUI entry point and settings widget for improved user experience
