+++ src/gui/settings/settings_widget.py
@@ -1,24 +1,22 @@
 from typing import Any
-import os
 
 from PySide6.QtWidgets import (
     QWidget, 
     QVBoxLayout, 
     QLabel, 
-    QLineEdit, 
     QHBoxLayout, 
     QCheckBox,
     QComboBox, 
-    QPushButton, 
-    QDialog, 
+    QPushButton,
     QScrollArea
 )
-from PySide6.QtGui import QFont, QShortcut, QKeySequence, QMouseEvent, QIcon
-from PySide6.QtCore import Qt, QTimer, QObject, QEvent, QPoint, QSize
+from PySide6.QtGui import QFont
+from PySide6.QtCore import Qt, QTimer
 
 from controller.app_controller import AppController
 from gui.dialogs_user import show_dialog_type
 from gui.manage_widgets import ManagerWidgets
+from gui.settings.settings_elixirs_widget import SettingsElixirsWidget
 from config.config import res_abs_paths, scroll_bar_style
 
 class SettingsWidget(QWidget):
@@ -29,17 +27,13 @@ class SettingsWidget(QWidget):
     """
     def __init__(self):
         super().__init__()
-
         self.controller = AppController.get_instance()
-        self.matches_dialog = None  # Initialize matches_dialog attribute
 
         settings_data = self.controller.get_all_settings_data()
         if settings_data is None:
             show_dialog_type("Failed to load settings data. Please check the settings file.", "Settings load", "error", "no_action")
             QTimer.singleShot(0, lambda: ManagerWidgets.get_instance().set_page("home")) # Gives time to render actual widget before switching inmediately (if not it will not render main widget)
             return
-        
-        self.installEventFilter(self) # Install event filter to capture mouse events
 
         layout_main = QVBoxLayout(self)
         layout_main.setSpacing(60)
@@ -54,16 +48,6 @@ class SettingsWidget(QWidget):
             padding: 10px;
             border-radius: 10px;
         """)
-        self.elixir_labels_style = """
-            QLabel {
-                color: white;
-                background-color: rgb(50, 50, 50);
-                border: 1px solid rgb(80, 80, 80);
-                padding: 5px;
-                border-radius: 8px;
-            }
-        """
-        self.elixirs_default_font = QFont("Arial", 12)
         
         settings_title_label.setFont(QFont("Arial", 24, QFont.Weight.Bold))
         settings_title_label.setAlignment(Qt.AlignmentFlag.AlignHCenter)
@@ -226,15 +210,9 @@ class SettingsWidget(QWidget):
                 self.scroll_area_elixirs.setFixedHeight(200)
                 self.scroll_area_elixirs.setFixedWidth(400)
 
-                elixirs_widget = QWidget()
-                self.elixirs_layout = QVBoxLayout(elixirs_widget)
-                self.elixirs_layout.setAlignment(Qt.AlignmentFlag.AlignTop)
-                self.elixirs_layout.setSpacing(5)
+                self.elixirs_widget = SettingsElixirsWidget(setting_val, self.settings_actual_data, self.on_settings_changed) # Create an instance of the elixirs widget
                 
-                self.scroll_area_elixirs.setWidget(elixirs_widget)
-
-                for id, name in setting_val.items():
-                    self.add_elixir_entry(name, id)
+                self.scroll_area_elixirs.setWidget(self.elixirs_widget)
 
                 setting_layout.addWidget(self.scroll_area_elixirs, 0, Qt.AlignmentFlag.AlignRight | Qt.AlignmentFlag.AlignVCenter)
                 layout_settings_inputs.addWidget(setting_widget) # Add the setting to the settings container widget
@@ -243,26 +221,10 @@ class SettingsWidget(QWidget):
                 search_widget = QWidget()
                 search_layout = QHBoxLayout(search_widget)
 
-                self.search_line_edit = QLineEdit()
-                self.search_line_edit.setFixedWidth(400)
-                self.search_line_edit.setPlaceholderText("Elixir Name or ID")
-                self.search_line_edit.setFont(QFont("Arial", 14))
-                self.search_line_edit.setClearButtonEnabled(True)
-                self.search_line_edit.setStyleSheet("""
-                    QLineEdit {
-                        background-color: rgb(50, 50, 50);
-                        color: white;
-                        border: 1px solid rgb(80, 80, 80);
-                        padding: 5px;
-                        border-radius: 8px;
-                    }
-                """)
+                search_elixir_line_edit = self.elixirs_widget.get_search_elixir_input()
+                search_elixir_line_edit.textChanged.connect(lambda text: self.elixirs_widget.search_elixir(text)) # type: ignore
 
-                self.search_line_edit.textChanged.connect(lambda text: self.search_elixir(text)) # type: ignore
-                self.esc_shortcut = QShortcut(QKeySequence("Escape"), self.search_line_edit)
-                self.esc_shortcut.activated.connect(lambda: self.matches_dialog.close() if self.matches_dialog else None)
-
-                search_layout.addWidget(self.search_line_edit, 0, Qt.AlignmentFlag.AlignRight | Qt.AlignmentFlag.AlignVCenter)
+                search_layout.addWidget(search_elixir_line_edit, 0, Qt.AlignmentFlag.AlignRight | Qt.AlignmentFlag.AlignVCenter)
                 layout_settings_inputs.addWidget(search_widget) # Add search widget to settings container widget
 
         layout_main.addWidget(widget_settings_inputs, 0, Qt.AlignmentFlag.AlignTop)
@@ -292,255 +254,9 @@ class SettingsWidget(QWidget):
                 background-color: rgb(150, 150, 150);
             }
         """)
-        self.apply_settings_button.clicked.connect(lambda _: self.save_user_settings()) # type: ignore
+        self.apply_settings_button.clicked.connect(lambda _: self.apply_user_settings()) # type: ignore
 
         layout_main.addWidget(self.apply_settings_button, 1, Qt.AlignmentFlag.AlignHCenter | Qt.AlignmentFlag.AlignBottom)
-
-    def add_elixir_entry(self, elixir_id: str, elixir_name: str):
-        """
-        Add an entry for an elixir in the settings widget.
-            :param elixir_id: The ID of the elixir.
-            :param elixir_name: The name of the elixir.
-        """
-        entry_elixir_widget = QWidget()
-        entry_elixir_layout = QHBoxLayout(entry_elixir_widget)
-        entry_elixir_layout.setContentsMargins(0, 2, 0, 2)
-        entry_elixir_layout.setSpacing(10)
-
-        button_delete_elixir = QPushButton()
-        button_delete_elixir.setFont(self.elixirs_default_font)
-        button_delete_elixir.setIcon(QIcon(res_abs_paths["delete_elixir"]) if os.path.exists(res_abs_paths["delete_elixir"]) else QIcon(res_abs_paths["not_found_ico"]))
-        button_delete_elixir.setStyleSheet("""
-            QPushButton {
-                background-color: transparent;
-                border-radius: 14px;
-                padding: 2px;
-            }
-            QPushButton:hover {
-                background-color: rgb(120, 60, 60);
-            }
-            QPushButton:pressed {
-                background-color: rgb(160, 60, 60);
-            }
-        """)
-        button_delete_elixir.setIconSize(QSize(25, 25))
-
-        button_delete_elixir.clicked.connect(lambda _, widget=entry_elixir_widget, name=elixir_name: self.delete_elixir_entry(widget, name)) # type: ignore
-
-        label_elixir = QLabel(f"{elixir_name} ({elixir_id})")
-        label_elixir.setFont(self.elixirs_default_font)
-        label_elixir.setStyleSheet(self.elixir_labels_style)
-
-        entry_elixir_layout.addWidget(button_delete_elixir, 0, Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
-        entry_elixir_layout.addWidget(label_elixir, 0, Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
-        self.elixirs_layout.addWidget(entry_elixir_widget, 0, Qt.AlignmentFlag.AlignLeft | Qt.AlignmentFlag.AlignVCenter)
-
-    def delete_elixir_entry(self, widget: QWidget, elixir_name: str):
-        """
-        Delete an elixir entry from the settings widget.
-            :param widget: The widget representing the elixir entry to be deleted.
-            :param elixir_name: The name of the elixir to be deleted.
-        """
-        self.elixirs_layout.removeWidget(widget)
-        widget.setParent(None)
-        widget.deleteLater()
-
-        self.settings_actual_data['Elixirs'][1].pop(elixir_name, None)  # Remove the elixir from the settings data
-        self.on_settings_changed('Elixirs', self.settings_actual_data['Elixirs'][1])  # Trigger settings change
-
-    def eventFilter(self, obj: QObject, event: QEvent) -> bool:
-        """
-        Event filter to handle mouse events on the settings widget.
-            :param obj: The object that received the event.
-            :param event: The event that occurred.
-            :return: True if the event was handled, False otherwise.
-        """
-        if event.type() == QEvent.Type.MouseButtonPress and self.matches_dialog and self.matches_dialog.isVisible():
-            if isinstance(event, QMouseEvent):
-                # Find the widget under the mouse click
-                clicked_widget = self.childAt(self.mapFromGlobal(event.globalPos()))
-                
-                # If clicked widget is None or NOT a child/descendant of matches_dialog, close dialog
-                if clicked_widget is None or not (clicked_widget == self.matches_dialog or self.matches_dialog.isAncestorOf(clicked_widget)):
-                    self.matches_dialog.close()
-                        
-        return super().eventFilter(obj, event)
-
-    def search_elixir(self, text: str):
-        """
-        Search for elixirs by name or ID and update the elixir settings.
-            :param text: The text to search for in the elixirs.
-        """
-        matches = self.controller.get_match_elixirs(text)
-        if matches is None:
-            self.clean_matches_dialog()  # Clean previous dialog if exists
-            return # Empty matches, do nothing
-        self.show_elixir_matches(matches)
-
-    def clean_matches_dialog(self):
-        """
-        Clean the matches dialog if it exists.
-        This is used to ensure that the dialog is not duplicated when searching for elixirs.
-        """
-        if self.matches_dialog:
-            self.matches_dialog.close()
-            self.matches_dialog = None
-
-    def show_elixir_matches(self, matches: dict[str, str] | str):
-        """
-        Show a dialog with the elixir matches found based on the search text.
-            :param matches: A dictionary of elixir matches found (id: name) or a string indicating no matches found.
-        """
-        self.clean_matches_dialog()  # Clean previous dialog if exists
-
-        # Create a dialog to show the elixir matches
-        self.matches_dialog = QDialog(self)
-        self.matches_dialog.setStyleSheet("""
-            QDialog {
-                border: 2px solid rgb(120, 120, 120);
-                border-radius: 10px;
-                background-color: rgb(30, 30, 30);
-                padding: 5px;
-            }
-        """)
-        
-        self.matches_dialog.setWindowFlags(
-            Qt.WindowType.FramelessWindowHint |
-            Qt.WindowType.WindowStaysOnTopHint
-        )
-
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        scroll_area.setStyleSheet(f"""
-            QScrollArea {{ 
-                background-color: transparent;
-            }}
-            
-            {scroll_bar_style}
-
-            QScrollBar::sub-line:vertical {{ /* Up arrow */
-                background: rgb(150, 150, 150);
-                height: 25px;
-                width: 25px;
-                subcontrol-position: top;
-                subcontrol-origin: margin;
-                border-radius: 5px;
-                image: url("{res_abs_paths['up_arrow']}");
-            }}
-
-            QScrollBar::add-line:vertical {{ /* Down arrow */
-                background: rgb(150, 150, 150);
-                height: 25px;
-                width: 25px;
-                subcontrol-position: bottom;
-                subcontrol-origin: margin;
-                border-radius: 5px;
-                image: url("{res_abs_paths['down_arrow']}");
-            }}
-
-            QScrollBar::sub-line:horizontal {{ /* Left arrow */
-                background: rgb(150, 150, 150);
-                height: 20px;
-                width: 18px;
-                subcontrol-position: left;
-                subcontrol-origin: margin;
-                border-radius: 5px;
-                image: url("{res_abs_paths['left_arrow']}");
-            }}
-            
-            QScrollBar::add-line:horizontal {{ /* Right arrow */
-                background: rgb(150, 150, 150);
-                height: 20px;
-                width: 18px;
-                subcontrol-position: right;
-                subcontrol-origin: margin;
-                border-radius: 5px;
-                image: url("{res_abs_paths['right_arrow']}");
-            }}
-        """)
-
-        content_widget = QWidget()
-        matches_layout = QVBoxLayout(content_widget)
-
-        if isinstance(matches, str):
-            # If matches is a string, it means no matches were found
-            no_matches_label = QLabel(matches)
-            no_matches_label.setStyleSheet("""
-                color: white;
-                font-size: 16px;
-                padding: 10px;
-            """)
-            no_matches_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-            matches_layout.addWidget(no_matches_label)
-
-        else:
-            for id, name in matches.items():
-                match_button = QPushButton(f"{name} ({id})")
-                match_button.setFont(QFont("Arial", 12))
-                match_button.setMinimumHeight(30)
-                match_button.setStyleSheet("""
-                    QPushButton {
-                        background-color: rgba(255, 255, 255, 0.2);
-                        border: 1px solid rgba(255, 255, 255, 0.5);
-                        border-radius: 6px;
-                        color: rgb(220, 220, 220);
-                    }
-                    QPushButton:hover {
-                        background-color: rgba(255, 255, 255, 0.5);
-                    }
-                    QPushButton:pressed {
-                        background-color: rgba(255, 255, 255, 0.7);
-                    }
-                    QToolTip {
-                        background-color: rgb(30, 30, 30);;
-                        border: 1px solid rgb(120, 120, 120);
-                        border-radius: 6px;
-                        font-size: 14px;
-                    }              
-                """)
-
-                match_button.clicked.connect(lambda _, elixir_name=name, elixir_id=id: self.update_elixirs_list(elixir_name, elixir_id)) # type: ignore
-                matches_layout.addWidget(match_button)
-
-        scroll_area.setWidget(content_widget)
-
-        dialog_layout = QVBoxLayout(self.matches_dialog)
-        dialog_layout.addWidget(scroll_area)
-
-        self.matches_dialog.adjustSize()
-        self.matches_dialog.setMaximumHeight(200)
-        self.matches_dialog.setMinimumWidth(500)
-
-        if self.search_line_edit:
-            height_search_line = self.search_line_edit.height()
-            pos = self.search_line_edit.mapTo(self, QPoint(0, height_search_line))
-            offset_x = (self.matches_dialog.width() - self.search_line_edit.width()) // 2  # Center the dialog horizontally below the search line edit 
-            offset_y = 20
-            self.matches_dialog.move(pos.x() - offset_x, pos.y() + offset_y)  # Position dialog below the search line edit
-        
-        self.matches_dialog.show() # Show dialog with matches
-
-    def update_elixirs_list(self, elixir_name: str, elixir_id: str):
-        """
-        Update the elixirs list with the selected elixir from the matches dialog.
-            :param elixir_name: The name of the elixir to add.
-            :param elixir_id: The ID of the elixir to add.
-        """
-        id_entry_json, elixirs_dict = self.settings_actual_data['Elixirs'] # Get the actual elixirs list
-
-        if elixir_id in elixirs_dict.values():
-            show_dialog_type(f"Elixir {elixir_name} ({elixir_id}) is already in the list.", "Add elixir", "info", "no_action")
-            return # If the elixir ID is already in the dict, do nothing
-
-        elixirs_dict[elixir_name] = elixir_id # Add the new elixir to the list
-        self.settings_actual_data['Elixirs'] = (id_entry_json, elixirs_dict)
-
-        self.add_elixir_entry(elixir_id, elixir_name) # Add the new elixir entry to the UI
-
-        self.search_line_edit.setText("") # Clear the search line edit to show the updated elixirs list
-        self.matches_dialog.close() if self.matches_dialog else None # Close the dialog with matches
-
-        self.on_settings_changed('Elixirs', elixirs_dict) # Notify that the settings have changed
 
     def update_original_settings(self):
         """
@@ -553,9 +269,13 @@ class SettingsWidget(QWidget):
             else:
                 self.original_settings[key] = value[1]
 
-    def save_user_settings(self):
+    def apply_user_settings(self):
+        """ Apply the user settings by saving the current actual data to the controller.
+            This method disables the widget while saving settings and re-enables it after saving.
+            If the settings are saved successfully, it updates the original settings to match the current actual data.
+        """
         self.setEnabled(False) # Disable the widget while saving settings
-        result = self.controller.save_user_settings(self.settings_actual_data)
+        result = self.controller.apply_user_settings(self.settings_actual_data)
 
         if result == 0: # If the settings were saved successfully
             self.update_original_settings() # Update the original settings to match the current actual data
